More progress on reactor.
