+++ server/classes/reactor.js
@@ -10,8 +10,22 @@ export default class Reactor extends System {
     this.type = params.type || 'reactor';
     this.powerOutput = params.powerOutput || 120;
     this.efficiency = params.efficiency || 1;
-    this.powerMode = params.powerMode || 'internal';
     this.batteryChargeLevel = params.batteryChargeLevel || 1;
-    this.batteryChargeRate = params.batteryChargeRate || 1/1000;
+    this.batteryChargeRate = params.batteryChargeRate || 1 / 1000;
+  }
+  eject(tf = true){
+    this.ejected = tf;
+  }
+  changeOutput(output){
+    this.powerOutput = output;
+  }
+  changeEfficiency(efficiency){
+    this.efficiency = efficiency;
+  }
+  changeBatteryChargeLevel(level){
+    this.batteryChargeLevel = level;
+  }
+  changeBatteryChargeRate(rate){
+    this.batteryChargeRate = rate;
   }
 }
