+++ server/events/flightStructure.js
@@ -1,7 +1,6 @@
 import App from '../../app.js';
 import { pubsub } from '../helpers/subscriptionManager.js';
 import * as Classes from '../classes';
-import uuid from 'uuid';
 
 function getTimelineObject(simulatorId, missionId) {
   let object;
@@ -117,37 +116,6 @@ App.on('changeSimulatorCrewCount', ({ simulatorId, crewCount }) => {
   }
   pubsub.publish('simulatorsUpdate', App.simulators);
 });
-App.on('shipDockingChanged', ({ simulatorId, which, state}) => {
-  const simulator = App.simulators.find(s => s.id === simulatorId);
-  if (simulator) {
-    simulator[which](state);
-  }
-  pubsub.publish('simulatorsUpdate', App.simulators);
-})
-App.on('remoteAccessSendCode', ({ simulatorId, code, station }) => {
-  const simulator = App.simulators.find(s => s.id === simulatorId);
-  if (simulator) {
-    simulator.sendCode(code, station);
-  }
-  pubsub.publish('simulatorsUpdate', App.simulators);
-})
-App.on('remoteAccessUpdateCode', ({ simulatorId, codeId, state }) => {
-  const simulator = App.simulators.find(s => s.id === simulatorId);
-  if (simulator) {
-    simulator.updateCode(codeId, state);
-  }
-  pubsub.publish('simulatorsUpdate', App.simulators);
-  //Send the notification
-  //Get the code
-  const code = simulator.ship.remoteAccessCodes.find(c => c.id === codeId);
-  pubsub.publish('notify', {id: uuid.v4(), 
-    simulatorId: simulator.id,
-    station: code.station,
-    title: 'Remote Access Code',
-    body: `Remote Access Code ${code.code} was ${state}`,
-    color: state === 'Accepted' ? 'success' : 'danger',
-  });
-})
 
 // Timeline
 App.on('addTimelineStep', ({ simulatorId, missionId, timelineStepId, name, description }) => {
