+++ server/events/index.js
@@ -12,3 +12,4 @@ import './internalComm';
 import './systems';
 import './navigation';
 import './shortRangeComm';
+import './ship';
