+++ server/processes/index.js
@@ -1,0 +1,5 @@
+import './engines';
+import './thrusters';
+import './sensorContacts';
+import './clientPing';
+import './reactor';
